chore: Rename Components folder to components #76
